+++ R/LearnerRegrXgboost.R
@@ -9,11 +9,12 @@
 #' To compute on GPUs, you first need to compile \CRANpkg{xgboost} yourself and link
 #' against CUDA. See \url{https://xgboost.readthedocs.io/en/stable/build.html#building-with-gpu-support}.
 #'
-#' Note that using the `watchlist` parameter directly will lead to problems when wrapping this [`Learner`] in a
+#' Note that using the `watchlist` parameter directly will lead to problems when wrapping this [mlr3::Learner] in a
 #' `mlr3pipelines` `GraphLearner` as the preprocessing steps will not be applied to the data in the watchlist.
+#' See the section *Early Stopping and Validation* on how to do this.
 #'
 #' @template note_xgboost
-#' @inheritSection mlr_learners_classif.xgboost Early stopping
+#' @inheritSection mlr_learners_classif.xgboost Early Stopping and Validation
 #' @inheritSection mlr_learners_classif.xgboost Initial parameter values
 #'
 #' @templateVar id regr.xgboost
@@ -31,19 +32,20 @@
 #' # Train learner with early stopping on spam data set
 #' task = tsk("mtcars")
 #'
-#' # Split task into training and test set
-#' split = partition(task, ratio = 0.8)
-#' task$set_row_roles(split$test, "test")
-#'
+#' # use 30 percent for validation
 #' # Set early stopping parameter
 #' learner = lrn("regr.xgboost",
 #'   nrounds = 100,
 #'   early_stopping_rounds = 10,
-#'   early_stopping_set = "test"
+#'   validate = 0.3
 #' )
 #'
 #' # Train learner with early stopping
 #' learner$train(task)
+#'
+#' # Inspect optimal nrounds and validation performance
+#' learner$internal_tuned_values
+#' learner$internal_valid_scores
 #' }
 LearnerRegrXgboost = R6Class("LearnerRegrXgboost",
   inherit = LearnerRegr,
@@ -52,6 +54,17 @@ LearnerRegrXgboost = R6Class("LearnerRegrXgboost",
     #' @description
     #' Creates a new instance of this [R6][R6::R6Class] class.
     initialize = function() {
+
+      p_nrounds = p_int(1L,
+        tags = c("train", "hotstart", "internal_tuning"),
+        aggr = crate(function(x) as.integer(ceiling(mean(unlist(x)))), .parent = topenv()),
+        in_tune_fn = crate(function(domain, param_vals) {
+          if (is.null(param_vals$early_stopping_rounds)) {
+            stop("Parameter 'early_stopping_rounds' must be set to use internal tuning.")
+          }
+          assert_integerish(domain$upper, len = 1L, any.missing = FALSE) }, .parent = topenv()),
+        disable_in_tune = list(early_stopping_rounds = NULL)
+      )
       ps = ps(
         alpha                       = p_dbl(0, default = 0, tags = "train"),
         approxcontrib               = p_lgl(default = FALSE, tags = "predict"),
@@ -64,7 +77,6 @@ LearnerRegrXgboost = R6Class("LearnerRegrXgboost",
         device                      = p_uty(default = "cpu", tags = "train"),
         disable_default_eval_metric = p_lgl(default = FALSE, tags = "train"),
         early_stopping_rounds       = p_int(1L, default = NULL, special_vals = list(NULL), tags = "train"),
-        early_stopping_set          = p_fct(c("none", "train", "test"), default = "none", tags = "train"),
         eta                         = p_dbl(0, 1, default = 0.3, tags = "train"),
         eval_metric                 = p_uty(default = "rmse", tags = "train"),
         feature_selector            = p_fct(c("cyclic", "shuffle", "random", "greedy", "thrifty"), default = "cyclic", tags = "train", depends = quote(booster == "gblinear")),
@@ -84,7 +96,7 @@ LearnerRegrXgboost = R6Class("LearnerRegrXgboost",
         missing                     = p_dbl(default = NA, tags = c("train", "predict"), special_vals = list(NA, NA_real_, NULL)),
         monotone_constraints        = p_uty(default = 0, tags = c("train", "control"), custom_check = crate(function(x) { checkmate::check_integerish(x, lower = -1, upper = 1, any.missing = FALSE) })), # nolint
         normalize_type              = p_fct(c("tree", "forest"), default = "tree", tags = "train", depends = quote(booster == "dart")),
-        nrounds                     = p_int(1L, tags = c("train", "hotstart")),
+        nrounds                     = p_nrounds,
         nthread                     = p_int(1L, default = 1L, tags = c("train", "threads")),
         ntreelimit                  = p_int(1, default = NULL, special_vals = list(NULL), tags = "predict"),
         num_parallel_tree           = p_int(1L, default = 1L, tags = "train"),
@@ -121,13 +133,13 @@ LearnerRegrXgboost = R6Class("LearnerRegrXgboost",
       # param deps
 
       # custom defaults
-      ps$values = list(nrounds = 1L, nthread = 1L, verbose = 0L, early_stopping_set = "none")
+      ps$values = list(nrounds = 1L, nthread = 1L, verbose = 0L)
 
       super$initialize(
         id = "regr.xgboost",
         param_set = ps,
         feature_types = c("logical", "integer", "numeric"),
-        properties = c("weights", "missings", "importance", "hotstart_forward"),
+        properties = c("weights", "missings", "importance", "hotstart_forward", "internal_tuning", "validation"),
         packages = c("mlr3learners", "xgboost"),
         label = "Extreme Gradient Boosting",
         man = "mlr3learners::mlr_learners_regr.xgboost"
@@ -173,7 +185,33 @@ LearnerRegrXgboost = R6Class("LearnerRegrXgboost",
     }
   ),
 
+  active = list(
+    #' @field internal_valid_scores (named `list()` or `NULL`)
+    #' The validation scores extracted from `model$evaluation_log`.
+    #' If early stopping is activated, this contains the validation scores of the model for the optimal `nrounds`,
+    #' otherwise the `nrounds` for the final model.
+    internal_valid_scores = function() {
+      self$state$internal_valid_scores
+    },
+    #' @field internal_tuned_values (named `list()` or `NULL`)
+    #' If early stopping is activated, this returns a list with `nrounds`,
+    #' which is extracted from `$best_iteration` of the model and otherwise `NULL`.
+    internal_tuned_values = function() {
+      self$state$internal_tuned_values
+    },
+    #' @field validate (`numeric(1)` or `character(1)` or `NULL`)
+    #' How to construct the internal validation data. This parameter can be either `NULL`,
+    #' a ratio, `"test"`, or `"predefined"`.
+    validate = function(rhs) {
+      if (!missing(rhs)) {
+        private$.validate = assert_validate(rhs)
+      }
+      private$.validate
+    }
+  ),
+
   private = list(
+    .validate = NULL,
     .train = function(task) {
 
       pv = self$param_set$get_values(tags = "train")
@@ -190,19 +228,18 @@ LearnerRegrXgboost = R6Class("LearnerRegrXgboost",
         xgboost::setinfo(data, "weight", task$weights$weight)
       }
 
-      if (pv$early_stopping_set != "none") {
-        pv$watchlist = c(pv$watchlist, list(train = data))
-      }
-
       # the last element in the watchlist is used as the early stopping set
 
-      if (pv$early_stopping_set == "test" && !is.null(task$row_roles$test)) {
+      internal_valid_task = task$internal_valid_task
+      if (!is.null(pv$early_stopping_rounds) && is.null(internal_valid_task)) {
+        stopf("Learner (%s): Configure field 'validate' to enable early stopping.", self$id)
+      }
+      if (!is.null(internal_valid_task)) {
         test_data = task$data(rows = task$row_roles$test, cols = task$feature_names)
         test_target =  task$data(rows = task$row_roles$test, cols = task$target_names)
         test_data = xgboost::xgb.DMatrix(data = as_numeric_matrix(test_data), label = data.matrix(test_target))
         pv$watchlist = c(pv$watchlist, list(test = test_data))
       }
-      pv$early_stopping_set = NULL
 
       if (!is.null(pv$holdout_task)) {
         holdout_task = pv$holdout_task$clone()
@@ -216,7 +253,7 @@ LearnerRegrXgboost = R6Class("LearnerRegrXgboost",
 
       invoke(xgboost::xgb.train, data = data, .args = pv)
     },
-
+    #' Returns the `$best_iteration` when early stopping is activated.
     .predict = function(task) {
       pv = self$param_set$get_values(tags = "predict")
       model = self$model
@@ -233,7 +270,6 @@ LearnerRegrXgboost = R6Class("LearnerRegrXgboost",
       if (!is.null(pars_train$early_stopping_rounds)) {
         stop("The parameter `early_stopping_rounds` is set. Early stopping and hotstarting are incompatible.")
       }
-      pars$early_stopping_set = NULL
 
       # Calculate additional boosting iterations
       # niter in model and nrounds in ps should be equal after train and continue
@@ -245,6 +281,25 @@ LearnerRegrXgboost = R6Class("LearnerRegrXgboost",
       data = xgboost::xgb.DMatrix(data = as_numeric_matrix(data), label = data.matrix(target))
 
       invoke(xgboost::xgb.train, data = data, xgb_model = model, .args = pars)
+    },
+
+    .extract_internal_tuned_values = function() {
+      if (is.null(self$state$param_vals$early_stopping_rounds)) {
+        return(NULL)
+      }
+      list(nrounds = self$model$best_iteration)
+    },
+
+    .extract_internal_valid_scores = function() {
+      if (is.null(self$model$evaluation_log)) {
+        return(named_list())
+      }
+      iter = if (!is.null(self$model$best_iteration)) self$model$best_iteration else self$model$niter
+      as.list(self$model$evaluation_log[
+        iter,
+        set_names(get(".SD"), gsub("^test_", "", colnames(get(".SD",)))),
+        .SDcols = patterns("^test_")
+      ])
     }
   )
 )